Wait for connect to finish sending all data
Goroutines created in connect, transfer data from a handler to telnet
client. However, as there's no synchronization between the goroutines
and the telnet server, there's a case where the server sends a prompt
before the goroutines finishsed all transmissions.

This fix assure that the telnet server waits until all data are sent to
the client, before sending a prompt.